move PaddedIt to its own module
